Backend/api/serializers.py
from rest_framework import serializers
from .models import *
class CategorySerializer(serializers.ModelSerializer):
    class Meta: 
        model=Category
        fields=[
            'id',
            'name',
            'slug',
            'views',
        ]

class PostSerializer(serializers.ModelSerializer):
    category=CategorySerializer()
    class Meta: 
        model=Post
        fields=[
            'id',
            'title',
            'slug',
            'thumbnail',
            'description',
            'content',
            'time_read',
            'published',
            'views',
            'category'
        ]

class PostListSerializer(serializers.ModelSerializer):
    category=CategorySerializer()
    class Meta: 
        model=Post
        fields=[
            'id',
            'title',
            'slug',
            'thumbnail',
            'description',
            'time_read',
            'published',
            'views',
            'category'
        ]
